08  - Streamlit and prototyping/requirements.txt
streamlit==1.29.0
pandas==2.2.0
matplotlib==3.8.2
python-dotenv==1.0.0
openai==1.12.0
requests==2.31.0
numpy
scikit-learn

